feat: consolidate profile into portal page
consolidates the separate /profile page into /portal for a unified artist
experience. profile editing is now the first section in the portal, followed
by track upload and track management.

changes:
- merge profile editing into /portal as first section
- update header: @{handle} now links to public profile /u/{handle}
- remove redundant /profile route (keep /profile/setup for onboarding)
- add profile settings form with artist name, bio, and avatar fields
- simplify navigation with single portal destination
